refactor: consolidate video schema definitions into src/schema/videos/convert
