+++ src/apps/compute/videos/index.ts
@@ -1,10 +1,10 @@
 import express, { Router } from 'express';
 import { validateRequest } from 'src/utils/validator';
 import { convertHandler } from './routes/convert';
-import { ConvertHandlerRequest, ConvertHandlerSchema } from './routes/convert/schema';
+import { ConvertHandlerRequest, convertHandlerSchema } from 'src/schema/videos/convert';
 
 const videosRouter: Router = express.Router();
 
-videosRouter.post('/convert-handler', validateRequest<ConvertHandlerRequest>(ConvertHandlerSchema), convertHandler);
+videosRouter.post('/convert-handler', validateRequest<ConvertHandlerRequest>(convertHandlerSchema), convertHandler);
 
 export { videosRouter };
