+++ src/schema/videos/convert/index.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest';
-import { ConvertSchema } from './schema';
+import { convertHandlerSchema, convertSchema } from './index';
 
-describe('ConvertSchema', () => {
+describe('convertSchema', () => {
   const validPayload = {
     body: {
       event: {
@@ -26,7 +26,7 @@ describe('ConvertSchema', () => {
   };
 
   it('should validate a correct payload', () => {
-    const result = ConvertSchema.safeParse(validPayload);
+    const result = convertSchema.safeParse(validPayload);
     expect(result.success).toBe(true);
   });
 
@@ -45,7 +45,7 @@ describe('ConvertSchema', () => {
           },
         },
       };
-      const result = ConvertSchema.safeParse(invalidPayload);
+      const result = convertSchema.safeParse(invalidPayload);
       expect(result.success).toBe(false);
     });
 
@@ -63,7 +63,7 @@ describe('ConvertSchema', () => {
           },
         },
       };
-      const result = ConvertSchema.safeParse(invalidPayload);
+      const result = convertSchema.safeParse(invalidPayload);
       expect(result.success).toBe(false);
     });
   });
@@ -82,7 +82,7 @@ describe('ConvertSchema', () => {
           },
         },
       };
-      const result = ConvertSchema.safeParse(invalidPayload);
+      const result = convertSchema.safeParse(invalidPayload);
       expect(result.success).toBe(false);
     });
   });
@@ -93,7 +93,7 @@ describe('ConvertSchema', () => {
         ...validPayload,
         headers: {},
       };
-      const result = ConvertSchema.safeParse(invalidPayload);
+      const result = convertSchema.safeParse(invalidPayload);
       expect(result.success).toBe(false);
     });
 
@@ -105,13 +105,13 @@ describe('ConvertSchema', () => {
           'x-extra-header': 'some-value',
         },
       };
-      const result = ConvertSchema.safeParse(payloadWithExtraHeaders);
+      const result = convertSchema.safeParse(payloadWithExtraHeaders);
       expect(result.success).toBe(true);
     });
   });
 
   it('should transform payload correctly', () => {
-    const result = ConvertSchema.safeParse(validPayload);
+    const result = convertSchema.safeParse(validPayload);
     expect(result.success).toBe(true);
     if (result.success) {
       expect(result.data).toEqual({
@@ -152,7 +152,7 @@ describe('ConvertSchema', () => {
       },
       headers: validPayload.headers,
     };
-    const result = ConvertSchema.safeParse(invalidPayload);
+    const result = convertSchema.safeParse(invalidPayload);
     expect(result.success).toBe(false);
     if (!result.success) {
       expect(result.error.issues.some(issue => issue.path.includes('id') || issue.path.includes('user_id'))).toBe(true);
@@ -173,7 +173,7 @@ describe('ConvertSchema', () => {
         },
       },
     };
-    const result = ConvertSchema.safeParse(youtubePayload);
+    const result = convertSchema.safeParse(youtubePayload);
     expect(result.success).toBe(true);
     if (result.success) {
       expect(result.data.event.data).toMatchObject({
@@ -197,7 +197,7 @@ describe('ConvertSchema', () => {
         },
       },
     };
-    const result = ConvertSchema.safeParse(hlsPayload);
+    const result = convertSchema.safeParse(hlsPayload);
     expect(result.success).toBe(true);
     if (result.success) {
       expect(result.data.event.data).toMatchObject({
@@ -205,5 +205,86 @@ describe('ConvertSchema', () => {
         platform: null,
       });
     }
+  });
+});
+
+describe('convertHandlerSchema', () => {
+  const validHeaders = {
+    'content-type': 'application/json',
+    'x-task-id': '223e4567-e89b-12d3-a456-426614174001',
+  };
+
+  const validMetadata = {
+    id: 'event-789',
+    spanId: 'span-abc',
+    traceId: 'trace-def',
+  };
+
+  const validData = {
+    id: '123e4567-e89b-12d3-a456-426614174000',
+    userId: '123e4567-e89b-12d3-a456-426614174001',
+    videoUrl: 'https://storage.example.com/video.mp4',
+  };
+
+  const createRequest = ({ data = validData, metadata = validMetadata, headers = validHeaders } = {}) => ({
+    body: { data, metadata },
+    headers,
+  });
+
+  describe('headers validation', () => {
+    it('should reject request with missing x-task-id headers', () => {
+      const { 'x-task-id': _, ...headersWithoutTaskId } = validHeaders;
+      // @ts-expect-error
+      const result = convertHandlerSchema.safeParse(createRequest({ headers: headersWithoutTaskId }));
+      expect(result.success).toBe(false);
+    });
+
+    it('should reject request with missing content-type headers', () => {
+      const { 'content-type': _, ...headersWithoutContentType } = validHeaders;
+      // @ts-expect-error
+      const result = convertHandlerSchema.safeParse(createRequest({ headers: headersWithoutContentType }));
+      expect(result.success).toBe(false);
+    });
+  });
+
+  it('should reject request with missing required fields', () => {
+    const { id: _, ...dataWithoutId } = validData;
+    // @ts-expect-error
+    const result = convertHandlerSchema.safeParse(createRequest({ data: dataWithoutId }));
+    expect(result.success).toBe(false);
+  });
+
+  it('should reject empty string values', () => {
+    const result = convertHandlerSchema.safeParse(createRequest({ data: { ...validData, id: '' } }));
+    expect(result.success).toBe(false);
+  });
+
+  it('should validate correct request structure', () => {
+    const validRequest = createRequest();
+    const result = convertHandlerSchema.safeParse(validRequest);
+    expect(result.success).toBe(true);
+  });
+
+  it('should reject non-UUID id', () => {
+    const result = convertHandlerSchema.safeParse(createRequest({ data: { ...validData, id: 'invalid-id' } }));
+    expect(result.success).toBe(false);
+  });
+
+  it('should reject non-HTTPS video URL', () => {
+    const result = convertHandlerSchema.safeParse(
+      createRequest({
+        data: { ...validData, videoUrl: 'http://storage.example.com/video.mp4' },
+      })
+    );
+    expect(result.success).toBe(false);
+  });
+
+  it('should reject non-video file URL', () => {
+    const result = convertHandlerSchema.safeParse(
+      createRequest({
+        data: { ...validData, videoUrl: 'https://storage.example.com/image.jpg' },
+      })
+    );
+    expect(result.success).toBe(false);
   });
 });
